Add functionality ro declare winner after 5 rounds
I'm awful at diving the program into smaller bits so my functions are quite lenghty and
they have more than one functionality--which I know isn't encouraged--so I'm working on it.
That aside, there is still some bug in the program that I'm too lazy to figure out and that
is still being able to run playRound() even though the game has ended. I will probably work
on that on later versions.